% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{build_query}
\alias{build_query}
\title{build_query}
\usage{
build_query(query)
}
\arguments{
\item{query}{a string, the graphql query}
}
\value{
a list
}
\description{
build_query
}
\examples{
my_query <- "{
        compendia {
            fullName,
            description
        }
}"
build_query(my_query)
}
