% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{create_module_bf}
\alias{create_module_bf}
\title{create a module based on provided biological features}
\usage{
create_module_bf(
  compendium = "vespucci",
  biofeaturesNames = NULL,
  normalization = "limma",
  rank = "magnitude",
  top_n = 50,
  useIds = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{biofeaturesNames}{A character vector (gene_names)}

\item{normalization}{A string - either 'limma' (default),'tpm' or legacy as normalization}

\item{rank}{A string ('magnitude' as default)}

\item{top_n}{A numeric - an integer for selecting the top ranked samplesets}

\item{useIds}{A logical (FALSE as default) - It allows using biofeatureIds
b}
}
\value{
A matrix - the module
}
\description{
create a module based on provided biological features
}
\examples{
\dontrun{
my_bf <- c("VIT_00s0246g00220","VIT_00s0332g00060","VIT_00s0332g00110")
tmp <- get_biofeature_id(id_In = my_bf, useIds = F)$id
my_mod <- create_module_bf(biofeaturesNames= tmp,
normalization = "limma", top_n = 15, useIds = T)
}
}
