% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{create_module_ss}
\alias{create_module_ss}
\title{create_module_ss}
\usage{
create_module_ss(
  compendium = "vespucci",
  samplesetNames = NULL,
  samplesetIds = NULL,
  version = "legacy",
  rank = "uncentered_correlation"
)
}
\arguments{
\item{compendium}{string - the selected compendium}

\item{samplesetNames}{a character vector - the sampleSets names}

\item{version}{string ('legacy' as default)}

\item{rank}{string ('magnitude' as default) - use \code{\link{get_ranking}}}
}
\value{
a data.frame - the module
}
\description{
create_module_ss
}
\examples{
my_ss <- c("E-MTAB-1514.US_30_2.ch1-vs-E-MTAB-1514.US_03_1.ch1","E-MTAB-1514.US_03_3.ch1-vs-E-MTAB-1514.US_03_1.ch1")
mod_ss <- create_module_ss(samplesetNames = my_ss)
}
