% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{enrich_module}
\alias{enrich_module}
\title{show the enrichment for ontology terms for both sampleSets and biofeatures}
\usage{
enrich_module(compendium = "vespucci", module = NULL, normalization = "limma")
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{module}{A matrix with valid rownames (biofeatureNames) and colnames (samplesetsNames)}

\item{normalization}{A string - either 'limma' (default),'tpm' or legacy as normalization}
}
\value{
a list with two data.frame (PlantOntology and GeneOntology)
}
\description{
show the enrichment for ontology terms for both sampleSets and biofeatures
}
\examples{
\dontrun{
gene_names <- c('VIT_00s0246g00220','VIT_00s0332g00060','VIT_00s0332g00110',
'VIT_00s0332g00160','VIT_00s0396g00010','VIT_00s0505g00030','VIT_00s0505g00060'
,'VIT_00s0873g00020','VIT_00s0904g00010')
module_1 <- create_module(biofeaturesNames=gene_names, normalization = "limma")
enrich_module(module = module_1, normalization = "limma")
}
}
