% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_annotation_triples}
\alias{get_annotation_triples}
\title{get_annotation_triples}
\usage{
get_annotation_triples(compendium = "vespucci", ids = NULL)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{ids}{A string - unique id of a biofeature, a sample, etc.}
}
\value{
A matrix with 3 columns
}
\description{
get_annotation_triples
}
\examples{
get_annotation_triples(ids = "U2FtcGxlVHlwZTozMjAw")
my_ids <- get_biofeature_id(name_In = c("VIT_00s0332g00160","VIT_00s0396g00010"
,"VIT_00s0505g00030"))$id
get_annotation_triples(ids = my_ids)
}
