% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_available_compendia}
\alias{get_available_compendia}
\title{Get all available compendia in COMPASS}
\usage{
get_available_compendia()
}
\value{
a list with info and versions
}
\description{
Get all available compendia in COMPASS
}
\examples{
get_available_compendia()
}
