% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_available_normalization}
\alias{get_available_normalization}
\title{get_available_normalization}
\usage{
get_available_normalization(compendium = "vespucci", version = "latest")
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{version}{a character(default 'latest')}
}
\value{
a data.frame
}
\description{
get_available_normalization
}
\examples{
get_available_normalization()
}
