% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_available_plot_methods}
\alias{get_available_plot_methods}
\title{show all available plot methods}
\usage{
get_available_plot_methods(compendium = "vespucci")
}
\arguments{
\item{compendium}{A string - the selected compendium}
}
\value{
A list with plot methods for distribution, heatmap and network
}
\description{
show all available plot methods
}
\examples{
get_available_plot_methods()
}
