% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_annotation_rdf}
\alias{get_biofeature_annotation_rdf}
\title{get_biofeature_annotation_rdf}
\usage{
get_biofeature_annotation_rdf(
  compendium = "vespucci",
  ids = "QmlvRmVhdHVyZVR5cGU6Mg=="
)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{ids}{unique biologicafeature annotation from \code{\link{get_biofeature_by_name()}}}
}
\value{
a data.frame
}
\description{
get_biofeature_annotation_rdf
}
\examples{
get_biofeature_annotation_rdf(ids="QmlvRmVhdHVyZVR5cGU6Mg==")
}
