% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_annotations}
\alias{get_biofeature_annotations}
\title{get_biofeature_annotations}
\usage{
get_biofeature_annotations(compendium = "vespucci", name = "VIT_00s0332g00060")
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{name}{a character the biofeature of interest}
}
\value{
a data.frame
}
\description{
get_biofeature_annotations
}
\examples{
get_biofeature_annotations(name = "VIT_00s0332g00060")
}
