% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_by_ann_term}
\alias{get_biofeature_by_ann_term}
\title{get all the biofeatures (genes here) which include the selected term within the annotations values}
\usage{
get_biofeature_by_ann_term(compendium = "vitis_vinifera", term = "GO:0006260")
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{term}{a character}
}
\value{
a character vector
}
\description{
get all the biofeatures (genes here) which include the selected term within the annotations values
}
