% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_id}
\alias{get_biofeature_id}
\title{get_biofeature_id}
\usage{
get_biofeature_id(compendium = "vespucci", name_In = NULL)
}
\arguments{
\item{compendium}{string - the selected compendium}

\item{name_In}{a character vector - the biofeature names}
}
\value{
a data.frame
}
\description{
get_biofeature_id
}
\examples{
my_genes <- c("VIT_00s0332g00110","VIT_00s0332g00160","VIT_00s0396g00010","VIT_00s0505g00030")
get_biofeature_id(name_In = my_genes)
}
