% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_info}
\alias{get_biofeature_info}
\title{get_biofeature_info}
\usage{
get_biofeature_info(
  compendium = "vitis_vinifera",
  name = "VIT_00s0332g00060",
  ontology = "Gene ontology"
)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{name}{a character the biofeature of interest}

\item{ontology}{a character}
}
\value{
a data.frame
}
\description{
get_biofeature_info
}
\examples{
get_biofeature_info()
}
