% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_name}
\alias{get_biofeature_name}
\title{get_biofeature_name}
\usage{
get_biofeature_name(compendium = "vespucci", id_In = NULL)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{id_In}{A vector of character strings - the biofeature id;
if NULL it returns all biofeature ids}
}
\value{
A data.frame with three columns: id, name, description
}
\description{
get_biofeature_name
}
\examples{
my_ids <- c("QmlvRmVhdHVyZVZhbHVlVHlwZToyNzQzOQ==","QmlvRmVhdHVyZVZhbHVlVHlwZToyNzg5Mg==")
get_biofeature_name(id_In = my_ids)
}
