% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_ranking}
\alias{get_biofeature_ranking}
\title{get_biofeature_ranking()}
\usage{
get_biofeature_ranking(
  compendium = "vespucci",
  samplesetNames = NULL,
  version = "legacy",
  rank = "uncentered_correlation",
  rankTarget = "biofeatures",
  top_n = 50
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{samplesetNames}{A vector of character strings (here sampleset names)}

\item{version}{A string ('legacy' as default)}

\item{rank}{A string ('magnitude' as default) - use \code{\link{get_ranking}}}

\item{rankTarget}{A string ('biofeature' as default)}

\item{top_n}{A numeric - an integer for selecting the top ranked samplesets}
}
\value{
A data.frame with three columns id, name, value
}
\description{
get_biofeature_ranking()
}
\examples{
my_ss <- c("GSM671720.ch1-vs-GSM671719.ch1","GSM671721.ch1-vs-GSM671719.ch1"
,"GSM671722.ch1-vs-GSM671719.ch1","GSM147672.ch1-vs-GSM147690.ch1")
get_biofeature_ranking(samplesetNames = my_ss, top_n = 10)

}
