% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_compass_version}
\alias{get_compass_version}
\title{get the current version of COMPASS}
\usage{
get_compass_version()
}
\value{
A string
}
\description{
get the current version of COMPASS
}
\examples{
get_compass_version()
}
