% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_compendia}
\alias{get_compendia}
\title{Get all available compendia}
\usage{
get_compendia()
}
\value{
a list with each element containing name, fullName, descritpion and normaliaziotion for all available compendia.
}
\description{
Get all available compendia
}
\examples{
get_compendia()
}
