% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_compendium_data_source}
\alias{get_compendium_data_source}
\title{Get compendium data sources}
\usage{
get_compendium_data_source(compendium = "vespucci")
}
\arguments{
\item{compendium}{A string - the selected compendium}
}
\value{
A data.frame
}
\description{
Get compendium data sources
}
\examples{
get_compendium_data_source()
}
