% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_ids_from_alias}
\alias{get_ids_from_alias}
\title{get_ids_from_alias}
\usage{
get_ids_from_alias(
  compendium = "vespucci",
  target = "biofeature",
  alias_names = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{target}{A string - either sample or biofeature}

\item{alias_names}{A character vector - the aliases}
}
\value{
A character vector - the ids corresponding to the provided aliases
}
\description{
get_ids_from_alias
}
\examples{
my_names <- c('Vv00s0125g00280','Vv00s0187g00140','Vv00s0246g00010',
'Vv00s0246g00080','Vv00s0438g00020','Vv00s0246g00200','Q7M2G6','B9S8R7','B8XIJ8','D7SZ93')
get_ids_from_alias(target = "biofeature", alias_names = my_names)
}
