% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_ontologies}
\alias{get_ontologies}
\title{Get all ontologies for}
\usage{
get_ontologies(compendium = "vitis_vinifera")
}
\arguments{
\item{compendium}{a character - the selected compendium}
}
\value{
a data.frame
}
\description{
Get all ontologies for
}
