% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_ontology_structure}
\alias{get_ontology_structure}
\title{get_ontology_structure}
\usage{
get_ontology_structure(compendium = "vespucci", name_In = "Gene ontology")
}
\arguments{
\item{compendium}{character - the selected compendium}

\item{name_In}{character - the name of the ontology/ies of interest}
}
\value{
a character
}
\description{
get_ontology_structure
}
\examples{
get_ontology_structure(name_In="Agronomy")
get_ontology_structure(name_In=c("Agronomy","Gene ontology"))
}
