% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_platform_information}
\alias{get_platform_information}
\title{Get information about all available platforms for the selected compendium,
use \code{\link{get_compendia}} to check all the available compendia}
\usage{
get_platform_information(compendium = "vespucci")
}
\arguments{
\item{compendium}{a character - the selected compendium}
}
\value{
a data.frame with five columns:
platformAccessId, platformName, description, dataSource, platformType
}
\description{
Get information about all available platforms for the selected compendium,
use \code{\link{get_compendia}} to check all the available compendia
}
