% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_platforms}
\alias{get_platforms}
\title{Get all available platforms for the selected compendium,
use \code{\link{get_compendia}} to check all the available compendia}
\usage{
get_platforms(compendium = "vitis_vinifera")
}
\arguments{
\item{compendium}{a character - the selected compendium}
}
\value{
a data.frame with 4 columns:
platformAccessId, platformName, description, dataSource, platformType
}
\description{
Get all available platforms for the selected compendium,
use \code{\link{get_compendia}} to check all the available compendia
}
\examples{
get_platforms()
}
