% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sample_annotation}
\alias{get_sample_annotation}
\title{get annotations for n samples from the selected compendium}
\usage{
get_sample_annotation(compendium = "vespucci", n = 10)
}
\arguments{
\item{compendium}{A string- the selected compendium}

\item{n}{an integer: number of sample to retrieve (default 10)}
}
\value{
a data.frame with three columns: sampleId sampleName annotation
}
\description{
get annotations for n samples from the selected compendium
}
\examples{
get_sample_annotation(n=25)
}
