% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sample_by_ann_term}
\alias{get_sample_by_ann_term}
\title{get samples related to a specific annotation term}
\usage{
get_sample_by_ann_term(
  compendium = "vitis_vinifera",
  term = "GROWTH.GREENHOUSE"
)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{term}{a character}
}
\value{
a data.frame
}
\description{
get samples related to a specific annotation term
}
