% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sample_info}
\alias{get_sample_info}
\title{get_sample_info}
\usage{
get_sample_info(compendium = "vespucci", sampleName = NULL)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{sampleName}{a character - GEO Sample accession id (GSM)}
}
\value{

}
\description{
get_sample_info
}
\examples{
get_sample_info(sampleName = "GSM287866.ch1")
}
