% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samples}
\alias{get_samples}
\title{Get all available samples for the selected compendium,
use \code{\link{get_compendia}} to check all the available compendia}
\usage{
get_samples(compendium = "vitis_vinifera")
}
\arguments{
\item{compendium}{a character - the selected compendium}
}
\value{
a data.frame
}
\description{
Get all available samples for the selected compendium,
use \code{\link{get_compendia}} to check all the available compendia
}
