% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samples_annotation_triples}
\alias{get_samples_annotation_triples}
\title{get_samples_annotation_triples}
\usage{
get_samples_annotation_triples(
  compendium = "vespucci",
  ids = "QmlvRmVhdHVyZVR5cGU6MQ=="
)
}
\arguments{
\item{compendium}{string - the selected compendium}

\item{ids}{string}
}
\value{

}
\description{
get_samples_annotation_triples
}
\examples{
get_samples_annotation_triples()
}
