% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samples_by_gse}
\alias{get_samples_by_gse}
\title{get samples id, name (GSM), description from experiment id (GSE)}
\usage{
get_samples_by_gse(compendium = "vespucci", experimentAccessId = NULL)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{experimentAccessId}{A string - GSE (GEO Series (experiment) access id)}
}
\value{
A data.frame with three columns: sampleId, sampleName, sampleDescription
}
\description{
get samples id, name (GSM), description from experiment id (GSE)
}
\examples{
get_samples_by_gse(experimentAccessId = "GSE98923")
}
