% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samples_by_sampleSet}
\alias{get_samples_by_sampleSet}
\title{Get all samples id for a selected sampleSet}
\usage{
get_samples_by_sampleSet(
  compendium = "vespucci",
  normalization = "tpm",
  sampleSet_In = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string - 'tpm' (as default), 'limma' or 'legacy'}

\item{sampleSet_In}{A string - a character vector of sampleSet ids}
}
\value{
A data.frame with three columns: sampleId, sampleName, sampleDescription
}
\description{
Get all samples id for a selected sampleSet
}
\examples{
\dontrun{
get_samples_by_sampleSet(sampleSet_In = c("U2FtcGxlU2V0VHlwZTo1Njg3","U2FtcGxlU2V0VHlwZTo1Njg4",
 "U2FtcGxlU2V0VHlwZTo1Njg5"), normalization = "limma")
}
}
