% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samples_info}
\alias{get_samples_info}
\title{get_samples_info}
\usage{
get_samples_info(
  compendium = "vespucci",
  normalization = "limma",
  id_In = NULL,
  useIds = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string ('limma' as default)}

\item{id_In}{A character vector of sample ids}

\item{useIds}{A logical (FALSE as default) - It allows using sampleIds}
}
\value{
A data.frame
}
\description{
get_samples_info
}
\examples{

\dontrun{
get_samples_info(id_In = c("U2FtcGxlU2V0VHlwZTo2NDE5","U2FtcGxlU2V0VHlwZTo2NDIx"),
useIds = TRUE)
get_samples_info(id_In = "GSM287866.ch1", useIds = FALSE)
}
}
