% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sampleset_by_sampleid}
\alias{get_sampleset_by_sampleid}
\title{get_sampleset_by_sampleid}
\usage{
get_sampleset_by_sampleid(
  compendium = "vespucci",
  normalization = "tpm",
  samples = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string ('legacy' as default)}

\item{samples}{A vector of character strings - samples id}
}
\value{
A data.frame with two columns: id and name
}
\description{
get_sampleset_by_sampleid
}
\examples{
my_ids=c("U2FtcGxlVHlwZToxMzM4","U2FtcGxlU2V0VHlwZToxMzM3")
get_sampleset_by_sampleid(samples = my_ids, normalization = "tpm")
}
