% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sampleset_id}
\alias{get_sampleset_id}
\title{get_sampleset_id}
\usage{
get_sampleset_id(
  compendium = "vespucci",
  normalization = "legacy",
  id_In = NULL,
  useIds = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string ('legacy' as default)}

\item{id_In}{A vector of character strings - either samplesetNames or sampleSetIds}

\item{useIds}{A logical (FALSE as default) - It allows using sampleSetIds}
}
\value{
A data.frame with two columns: id and name
}
\description{
get_sampleset_id
}
\examples{
my_ss=c("GSM671720.ch1-vs-GSM671719.ch1","GSM671721.ch1-vs-GSM671719.ch1"
,"GSM671722.ch1-vs-GSM671719.ch1","GSM147672.ch1-vs-GSM147690.ch1")
get_sampleset_id(id_In = my_ss, normalization = "legacy")
my_ids=c("U2FtcGxlU2V0VHlwZTo2NDE5","U2FtcGxlU2V0VHlwZToyMTg2OA==","U2FtcGxlU2V0VHlwZToyMTg2Nw==")
get_sampleset_id(id_In = my_ids, normalization = "limma", useIds = TRUE)
}
