% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sampleset_id_by_name}
\alias{get_sampleset_id_by_name}
\title{get_sampleset_id_by_name}
\usage{
get_sampleset_id_by_name(
  compendium = "vitis_vinifera",
  name = "GSM786264.ch1-vs-GSM786258.ch1"
)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{name}{a character the sampleset of interest}

\item{ontology}{a character}
}
\value{
a character
}
\description{
get_sampleset_id_by_name
}
\examples{
get_sampleset_id_by_name()
}
