% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samplesets_ranking}
\alias{get_samplesets_ranking}
\title{get_samplesets_ranking}
\usage{
get_samplesets_ranking(
  compendium = "vespucci",
  version = "legacy",
  rankTarget = "samplesets",
  rank = "magnitude",
  biofeaturesNames = NULL,
  biofeaturesIds = NULL,
  top_n = 50
)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{version}{a character ('legacy' as default)}

\item{rankTarget}{a character ('sampleset' as default)}

\item{rank}{a character ('magnitude' as default)}

\item{biofeatureNames}{a character vector (here gene_names)}

\item{biofeatureIds}{a character vector (here gene_ids)}
}
\value{
a list with four character vectors (id,name,type,value)
}
\description{
get_samplesets_ranking
}
\examples{
my_ids <- c("VIT_00s0332g00110","VIT_00s0332g00160","VIT_00s0396g00010","VIT_00s0505g00030")
get_samplesets_ranking(biofeaturesNames = my_ids)
get_samplesets_ranking(biofeaturesIds=c("QmlvRmVhdHVyZVR5cGU6MQ==","QmlvRmVhdHVyZVR5cGU6Mg==", "QmlvRmVhdHVyZVR5cGU6Mw==","QmlvRmVhdHVyZVR5cGU6NA==","QmlvRmVhdHVyZVR5cGU6NQ=="))
}
