% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sparql_annotation_triples}
\alias{get_sparql_annotation_triples}
\title{get_sparql_annotation_triples}
\usage{
get_sparql_annotation_triples(
  compendium = "vespucci",
  target = "sample",
 
    query = "SELECT ?s ?p ?o WHERE { ?s <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://purl.obolibrary.org/obo/NCIT_C19157> . ?s <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://purl.obolibrary.org/obo/PO_0009010>}"
)
}
\arguments{
\item{compendium}{A string- the selected compendium}

\item{ids}{string}
}
\value{

}
\description{
get_sparql_annotation_triples
}
\examples{
get_sparql_annotation_triples()
}
