% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{merge_modules}
\alias{merge_modules}
\title{Union of two modules}
\usage{
merge_modules(mod1, mod2)
}
\arguments{
\item{mod2}{}
}
\value{

}
\description{
Union of two modules
}
\examples{
#mod1 <- create_module_ss()
#mod2 <- create_module_bf()
#mod12_union <- merge_modules(mod1,mod2)
}
