% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{plot_annotation_network}
\alias{plot_annotation_network}
\title{plot rdf network}
\usage{
plot_annotation_network(compendium = "vespucci", ids = NULL, viewer = FALSE)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{ids}{A string - unique id of a biofeature, a sample, etc.}

\item{viewer}{A logical - if TRUE will plot the html widget}
}
\value{
Either a json, an html or a plotly htmlwidget
}
\description{
plot rdf network
}
\examples{
\dontrun{
my_ids <- get_biofeature_id(name_In = c("VIT_00s0332g00160","VIT_00s0396g00010"
,"VIT_00s0505g00030"))$id
plot_annotation_network(ids = my_ids, viewer = TRUE)
}
}
