% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution_coexpressed_samplesets}
\alias{plot_distribution_coexpressed_samplesets}
\title{plot_heatmap}
\usage{
plot_distribution_coexpressed_samplesets(
  compendium = "vespucci",
  version = "legacy",
  biofeaturesNames = NULL,
  biofeaturesIds = NULL
)
}
\arguments{
\item{compendium}{string - the selected compendium}

\item{version}{string ('legacy' as default)}

\item{biofeaturesIds}{a character vector - the biofeature ids}

\item{biofeatureNames}{a character vector (here gene_names)}
}
\value{
an html page with the distribution
}
\description{
plot_heatmap
}
