% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  compendium = "vespucci",
  version = "legacy",
  biofeaturesNames = NULL,
  samplesetNames = NULL
)
}
\arguments{
\item{compendium}{string - the selected compendium}

\item{version}{string ('legacy' as default)}

\item{samplesetNames}{a character vector - the sampleSets names}

\item{biofeatureNames}{a character vector (here gene_names)}
}
\value{
an html page with the heatmap
}
\description{
plot_heatmap
}
