% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_module}
\alias{plot_module}
\title{plot_module}
\usage{
plot_module(
  compendium = "vespucci",
  plot_type = "heatmap",
  normalization = NULL,
  biofeaturesNames = NULL,
  samplesetNames = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{plot_type}{A string - the type of plot either 'heatmap', 'distributions' or 'network'}

\item{normalization}{A string - either 'limma','tpm_sample' or legacy as normalization}

\item{biofeaturesNames}{A character vector (here gene_names)}

\item{samplesetNames}{A character vector - the sampleSets names}
}
\value{
An html page with the heatmap
}
\description{
plot_module
}
\examples{
\dontrun{
gene_names <- c('VIT_00s0246g00220','VIT_00s0332g00060','VIT_00s0332g00110',
'VIT_00s0332g00160','VIT_00s0396g00010','VIT_00s0505g00030',
'VIT_00s0505g00060','VIT_00s0873g00020','VIT_00s0904g00010')
mod_bf <- create_module(biofeaturesNames = gene_names)
my_plot_html <- plot_heatmap(plot_type = "heatmap",
biofeaturesNames = gene_names, normalization = "legacy")
tempDir <- tempfile()
dir.create(tempDir)
htmlFile <- file.path(tempDir, "plot_heatmap.html")
xml2::write_html(my_plot_html,file=htmlFile)
rstudioapi::viewer(htmlFile)
}
}
