% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_module_distribution}
\alias{plot_module_distribution}
\title{Plot a distribution from a model}
\usage{
plot_module_distribution(
  compendium = "vespucci",
  module = NULL,
  normalization = "legacy",
  type = "json",
  plot = TRUE,
  plotType = "biological_features_uncentered_correlation_distribution",
  getRank = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{module}{A matrix with valid rownames (biofeatureNames) and colnames (samplesetsNames)}

\item{normalization}{A string - either 'limma','tpm' or legacy as normalization}

\item{type}{A string -  either 'html'  or 'json}

\item{plot}{A logical - it return the graphics object}

\item{plotType}{A string - see \code{\link{get_available_plot_methods}}}

\item{getRank}{A logical - if TRUE return the ranking}
}
\value{
Either a json, an html, a plotly htmlwidget or a data.frame with the ranking
}
\description{
Plot a distribution from a model
}
\examples{
\dontrun{
gene_names <- c('VIT_00s0246g00220','VIT_00s0332g00060','VIT_00s0332g00110',
'VIT_00s0332g00160','VIT_00s0396g00010','VIT_00s0505g00030','VIT_00s0505g00060',
'VIT_00s0873g00020','VIT_00s0904g00010')
module_1 <- create_module(biofeaturesNames=gene_names, normalization = "legacy")
plot_module_distribution(module = module_1,
plotType = "biological_features_uncentered_correlation_distribution", plot = TRUE)
plot_module_distribution(module = module_1,
plotType = "sample_sets_magnitude_distribution", plot = TRUE)
plot_module_distribution(module = module_1,
plotType = "sample_sets_coexpression_distribution", plot = TRUE)
}
}
