% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_module_heatmap}
\alias{plot_module_heatmap}
\title{Plot a heatmap from a model}
\usage{
plot_module_heatmap(
  compendium = "vespucci",
  module = NULL,
  normalization = "legacy",
  type = "json",
  plot = TRUE,
  sorted = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{module}{A matrix with valid rownames (biofeatureNames) and colnames (samplesetsNames)}

\item{normalization}{A string - either 'limma','tpm_sample' or legacy as normalization}

\item{type}{A string -  either 'html'  or 'json}

\item{plot}{A logical - it returns the graphics object}

\item{sorted}{A logical - it returns sorted index for both bf and ss}
}
\value{
Either a json, an html, a plotly htmlwidget, or a list of sorted features
}
\description{
Plot a heatmap from a model
}
\examples{
\dontrun{
gene_names <- c('VIT_00s0246g00220','VIT_00s0332g00060','VIT_00s0332g00110')
module_1 <- create_module(biofeaturesNames=gene_names, version = "legacy")
plot_module_heatmap(module = module_1, plot = TRUE)
plot_module_heatmap(module = module_1, sorted = TRUE, plot = FALSE)
}
}
