% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_module_network}
\alias{plot_module_network}
\title{Plot a network from a model}
\usage{
plot_module_network(
  compendium = "vespucci",
  module = NULL,
  normalization = "legacy",
  type = "json",
  plot = TRUE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{module}{A matrix with valid rownames (biofeatureNames) and colnames (samplesetsNames)}

\item{normalization}{A string - either 'limma','tpm_sample' or legacy as normalization}

\item{type}{A string -  either 'html'  or 'json}

\item{plot}{A logical - it return the graphics object}
}
\value{
Either a json, an html or a plotly htmlwidget
}
\description{
Plot a network from a model
}
\examples{
\dontrun{
gene_names <- c('VIT_00s0246g00220','VIT_00s0332g00060','VIT_00s0332g00110',
'VIT_00s0332g00160','VIT_00s0396g00010','VIT_00s0505g00030','VIT_00s0505g00060'
,'VIT_00s0873g00020','VIT_00s0904g00010')
module_1 <- create_module(biofeaturesNames=gene_names, version = "legacy")
plot_module_network(module = module_1, plot = FALSE)
}
}
