% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_network_coexpression}
\alias{plot_network_coexpression}
\title{plot_network_coexpression}
\usage{
plot_network_coexpression(
  compendium = "vespucci",
  normalization = "legacy",
  biofeaturesNames = NULL,
  samplesetNames = NULL,
  useIds = TRUE,
  type = "json"
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string - either 'limma','tpm_sample' or legacy as normalization}

\item{biofeaturesNames}{A character vector (here gene_names)}

\item{samplesetNames}{A character vector - the sampleSets names}

\item{useIds}{A logical - TRUE as default}

\item{type}{string -  either 'html'  or 'json}
}
\value{
Either a json, an html or a plotly htmlwidget
}
\description{
plot_network_coexpression
}
