% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDistribution}
\alias{plotDistribution}
\title{plotDistribution}
\usage{
plotDistribution(type = "html")
}
\arguments{
\item{type}{either 'html' (default) or 'json}
}
\value{
An html - the plot
}
\description{
plotDistribution
}
\examples{
\dontrun{
my_plot_html <- plotDistribution()
tempDir <- tempfile()
dir.create(tempDir)
htmlFile <- file.path(tempDir, "plotDistribution.html")
xml2::write_html(my_plot_html,file=htmlFile)
rstudioapi::viewer(htmlFile)
}
}
