% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompass-package.R
\docType{package}
\name{rcompass-package}
\alias{rcompass-package}
\alias{rcompass}
\title{COMPASS  (COMpendia Programmatic Access Support Software) API client library}
\description{
a package connecting R to COMPASS GraphQL API
}
\author{
Paolo Sonego \email{paolo.sonego@fmach.it}
}
