% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{totalCounts}
\alias{totalCounts}
\title{totalCounts for aggregated types}
\usage{
totalCounts(
  compendium = "vespucci",
  version = "legacy",
  aggregate_type = "biofeatures"
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{version}{A string - either (default) 'legacy' or '2.0'}

\item{aggregate_type}{A string - an aggregate type (either 'biofeatures','sampleSets',
'platforms', 'samples', 'experiments')}
}
\value{
A numeric - the total count for the selected aggregate type
}
\description{
totalCounts for aggregated types
}
\examples{
\dontrun{
totalCounts(version="1.0", aggregate_type="sampleSets")
totalCounts(version="1.0", aggregate_type="biofeatures")
totalCounts(aggregate_type = "platforms")
totalCounts(aggregate_type = "samples")
totalCounts(aggregate_type = "experiments")
}
}
