% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{view_plot}
\alias{view_plot}
\title{Show a plot using the rstudio viewer}
\usage{
view_plot(
  module = NULL,
  plotType = "biological_features_uncentered_correlation_distribution",
  normalization = "legacy",
  alternativeColoring = TRUE,
  threshold = 0.7,
  min = -6,
  max = 6
)
}
\arguments{
\item{module}{A matrix with valid rownames (biofeatureNames) and colnames (samplesetsNames)}

\item{plotType}{A string - see \code{\link{get_available_plot_methods}}}

\item{normalization}{A string - either 'limma','tpm' or legacy as normalization}

\item{alternativeColoring}{A logical - if TRUE (defautl) a color blind friendly palette is used}

\item{threshold}{A numeric - A Pearson correalation value}

\item{min}{A numeric (-6 default)}

\item{max}{A numeric (6 default)}
}
\value{
A plotly htmlwidget
}
\description{
Show a plot using the rstudio viewer
}
