% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{build_query}
\alias{build_query}
\title{Process a GraphQL query from COMPASS API}
\usage{
build_query(query)
}
\arguments{
\item{query}{A string, the GraphQL query}
}
\value{
A list
}
\description{
Process a GraphQL query from COMPASS API
}
\examples{
my_query <- "{
        compendia {
            fullName,
            description
        }
}"
build_query(my_query)
}
