% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{create_module}
\alias{create_module}
\title{create module with biological features and sample sets}
\usage{
create_module(
  compendium = "vespucci",
  biofeatureNames = NULL,
  samplesetNames = NULL,
  version = "legacy"
)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{biofeatureNames}{a character vector (here gene_names)}

\item{samplesetNames}{a character vector (here sampleset names)}

\item{version}{a character ('legacy' as default)}
}
\value{
a list with three data.frame
}
\description{
create module with biological features and sample sets
}
\examples{
\dontrun{
my_mod <- create_module(biofeatureNames=c("VIT_00s0332g00110","VIT_00s0332g00160",
"VIT_00s0396g00010","VIT_00s0505g00030"),
samplesetNames=c("GSE27180_48hours-1-vs-GSE27180_0h-2",
"E-MTAB-1514.US_03_3.ch1-vs-E-MTAB-1514.US_03_1.ch1"))
pheatmap::pheatmap(my_mod,col = RcolorBrewer::brewer.pal(11,name="RdBu"))
# compare to vespucci:
# url <- "http://vespucci.colombos.fmach.it/cws_data/export_data/colombos_20190828_nEwFDG.txt"
# vesp_test=readr::read_tsv(url)
}

}
