% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{create_module}
\alias{create_module}
\title{create moduel with biological features and sample sets}
\usage{
create_module(compendium = "vitis_vinifera",
  biofeaturesIds = c("QmlvRmVhdHVyZVR5cGU6NTIzMjU=",
  "QmlvRmVhdHVyZVR5cGU6NTIzMjY="),
  samplesetIds = c("U2FtcGxlU2V0VHlwZToxMjYw",
  "U2FtcGxlU2V0VHlwZToxMjYx", "U2FtcGxlU2V0VHlwZToxMjYy"))
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{biofeaturesIds}{a character vector}

\item{samplesetIds}{a character vector}
}
\value{
a data.frame
}
\description{
create moduel with biological features and sample sets
}
\examples{
create_module()
}
