% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{create_module}
\alias{create_module}
\title{create module providing both biological features and sample sets}
\usage{
create_module(
  compendium = "vespucci",
  version = "legacy",
  biofeaturesNames = NULL,
  samplesetNames = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{version}{A string ('legacy' as default)}

\item{biofeaturesNames}{A character vector (gene_names)}

\item{samplesetNames}{A character vector (sampleset names)}
}
\value{
A data.frame - the module
}
\description{
create module providing both biological features and sample sets
}
\examples{
\dontrun{
my_bf <- c("VIT_00s0332g00110","VIT_00s0332g00160","VIT_00s0396g00010","VIT_00s0505g00030")
my_ss <- c("GSM671720.ch1-vs-GSM671719.ch1","GSM671721.ch1-vs-GSM671719.ch1"
,"GSM671722.ch1-vs-GSM671719.ch1","GSM147672.ch1-vs-GSM147690.ch1")
my_mod <- create_module(biofeaturesNames = my_bf, samplesetNames = my_ss)
pheatmap::pheatmap(my_mod, col = RColorBrewer::brewer.pal(11,name="RdBu"))
# compare to vespucci:
# url <- "http://vespucci.colombos.fmach.it/cws_data/export_data/colombos_20190828_nEwFDG.txt"
# vesp_test=readr::read_tsv(url)
}

}
