% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{create_module_bf}
\alias{create_module_bf}
\title{create_module_bf}
\usage{
create_module_bf(
  compendium = "vespucci",
  biofeaturesNames = c("VIT_00s0332g00110", "VIT_00s0332g00160", "VIT_00s0396g00010",
    "VIT_00s0505g00030", "VIT_00s0505g00060", "VIT_00s0873g00020", "VIT_00s0904g00010"),
  version = "legacy",
  rank = "magnitude",
  rankTarget = "samplesets"
)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{rank}{a character ('magnitude' as default)}

\item{biofeatureNames}{a character vector (here gene_names)}

\item{normalization}{a character}
}
\value{
a data.frame
}
\description{
create_module_bf
}
\examples{
create_module_bf(biofeatureNames=c("VIT_00s0332g00110","VIT_00s0332g00160","VIT_00s0396g00010","VIT_00s0505g00030","VIT_00s0505g00060","VIT_00s0873g00020","VIT_00s0904g00010")
}
