% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{create_module_bf}
\alias{create_module_bf}
\title{create a module based on provided biological features}
\usage{
create_module_bf(
  compendium = "vespucci",
  biofeaturesNames = NULL,
  normalization = "legacy",
  rank = "magnitude",
  useIds = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{biofeaturesNames}{A character vector (gene_names)}

\item{normalization}{A string ('legacy' as default)}

\item{rank}{A string ('magnitude' as default)}

\item{useIds}{A logical (FALSE as default) - It allows using biofeatureIds
b}
}
\value{
A matrix - the module
}
\description{
create a module based on provided biological features
}
\examples{
\dontrun{
my_bf <- c("VIT_00s0246g00220","VIT_00s0332g00060","VIT_00s0332g00110"
,"VIT_00s0332g00160","VIT_00s0396g00010","VIT_00s0505g00030","VIT_00s0505g00060"
,"VIT_00s0873g00020","VIT_00s0904g00010")
mod_bf <- create_module_bf(biofeaturesNames=my_bf, normalization = "legacy")
}
}
