% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{create_module_ss}
\alias{create_module_ss}
\title{create_module_ss}
\usage{
create_module_ss(
  compendium = "vespucci",
  samplesetNames = NULL,
  version = "legacy",
  rank = "uncentered_correlation"
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{samplesetNames}{A character vector (sampleset names)}

\item{version}{A string ('legacy' as default)}

\item{rank}{A string ('magnitude' as default) - use \code{\link{get_ranking}}}
}
\value{
A data.frame - the module
}
\description{
create_module_ss
}
\examples{
my_ss <- c("GSM671720.ch1-vs-GSM671719.ch1","GSM671721.ch1-vs-GSM671719.ch1"
,"GSM671722.ch1-vs-GSM671719.ch1","GSM147672.ch1-vs-GSM147690.ch1")
mod_ss <- create_module_ss(samplesetNames = my_ss)
}
