% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_annotation_triples}
\alias{get_annotation_triples}
\title{get_annotation_triples}
\usage{
get_annotation_triples(
  compendium = "vespucci",
  biofeaturesNames = NULL,
  samplesetNames = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{biofeaturesNames}{A character vector (gene_names)}

\item{samplesetNames}{A character vector (sampleset names)}
}
\value{
A character vector
}
\description{
get_annotation_triples
}
\examples{
#get_annotation_triples(samplesetName = "GSM147672.ch1-vs-GSM147690.ch1") # TO FIX!
get_annotation_triples(biofeaturesNames = "VIT_00s0332g00110")
}
