% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_available_compendia}
\alias{get_available_compendia}
\title{Get all available compendia}
\usage{
get_available_compendia()
}
\value{
a list with each element containing name, fullName, description and normalization for all available compendia.
}
\description{
Get all available compendia
}
