% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_available_normalization}
\alias{get_available_normalization}
\title{get_available_normalization}
\usage{
get_available_normalization(compendium = "vespucci", version = "latest")
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{version}{A string - either (default) 'latest' or 'legacy'}
}
\value{
A vector of character strings with the available normalization methods
}
\description{
get_available_normalization
}
\examples{
get_available_normalization(version = "legacy")
}
