% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_annotations}
\alias{get_biofeature_annotations}
\title{get_biofeature_annotations}
\usage{
get_biofeature_annotations(compendium = "vespucci", bioFeature_Name_In = NULL)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{bioFeature_Name_In}{A character vector with bioFeature names}
}
\value{
A data.frame
}
\description{
get_biofeature_annotations
}
\examples{
get_biofeature_annotations(bioFeature_Name_In = c("VIT_00s0332g00060","VIT_00s0332g00070"))
}
