% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_by_name}
\alias{get_biofeature_by_name}
\title{get biofeature ids by name}
\usage{
get_biofeature_by_name(compendium = "vitis_vinifera",
  name = "VIT_00s0332g00060")
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{name}{a character - the biofeature name}
}
\value{
a character vector
}
\description{
get biofeature ids by name
}
\examples{
get_biofeature_by_name()
}
