% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_by_name}
\alias{get_biofeature_by_name}
\title{get biofeature sequence by name}
\usage{
get_biofeature_by_name(
  compendium = "vespucci",
  name_In = NULL,
  field = "sequence"
)
}
\arguments{
\item{compendium}{A string- the selected compendium}

\item{field}{the biofeature field of interest ('sequence' as default)}

\item{name}{A string- the biofeature (here gene) name}
}
\value{
a data.frame with three columns: name,id and value
}
\description{
get biofeature sequence by name
}
\examples{
get_biofeature_by_name(name_In = "VIT_00s0332g00060")
get_biofeature_by_name(name_In = c("VIT_00s0332g00060", "VIT_00s0246g00220", "VIT_00s0332g00160"))
}
