% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_id}
\alias{get_biofeature_id}
\title{get_biofeature_id}
\usage{
get_biofeature_id(compendium = "vitis_vinifera", name = "VIT_00s0332g00060")
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{name}{a character - the biofeature name}

\item{field}{the biofeature field of interest  (sequence as default)}
}
\value{
a list
}
\description{
get_biofeature_id
}
\examples{
get_biofeature_id()
}
