% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_id}
\alias{get_biofeature_id}
\title{get_biofeature_id}
\usage{
get_biofeature_id(compendium = "vespucci", id_In = NULL, useIds = TRUE)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{id_In}{A vector of character strings - the biofeature id;
if NULL it returns all biofeature ids}

\item{useIds}{A logical (TRUE as default) - It allows using biofeature ids}
}
\value{
A data.frame with three columns: id, name, description
}
\description{
get_biofeature_id
}
\examples{
my_genes <- c("VIT_00s0332g00110","VIT_00s0332g00160","VIT_00s0396g00010","VIT_00s0505g00030")
get_biofeature_id(id_In = my_genes, useIds = FALSE)
my_ids <- c("QmlvRmVhdHVyZVZhbHVlVHlwZToyNzQzOQ==","QmlvRmVhdHVyZVZhbHVlVHlwZToyNzg5Mg==")
get_biofeature_id(id_In = my_ids, useIds = TRUE)
}
