% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_ranking}
\alias{get_biofeature_ranking}
\title{Title}
\usage{
get_biofeature_ranking(
  compendium = "vespucci",
  samplesetIds = c("U2FtcGxlU2V0VHlwZToxMjYy", "U2FtcGxlU2V0VHlwZToxMjYw"),
  samplesetNames = NULL,
  version = "legacy",
  rank = "std"
)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{samplesetIds}{a character vector (here gene_names)}

\item{samplesetNames}{a character vector (here sampleset names)}

\item{version}{a character ('legacy' as default)}

\item{rank}{a character ('magnitude' as default) - use \code{\link{get_ranking()$sampleSets}}
for the available values}
}
\value{
a list with four character vectors (id,name,type,value)
}
\description{
Title
}
\examples{
get_biofeature_ranking()
}
