% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_ranking}
\alias{get_biofeature_ranking}
\title{get_biofeature_ranking()}
\usage{
get_biofeature_ranking(
  compendium = "vespucci",
  samplesetIds = NULL,
  samplesetNames = NULL,
  version = "legacy",
  rank = "uncentered_correlation",
  rankTarget = "biofeatures",
  top_n = 50
)
}
\arguments{
\item{compendium}{string - the selected compendium}

\item{samplesetIds}{a character vector (here gene_names)}

\item{samplesetNames}{a character vector (here sampleset names)}

\item{version}{string ('legacy' as default)}

\item{rank}{string ('magnitude' as default) - use \code{\link{get_ranking}}}

\item{rankTarget}{string ('biofeature' as default)
for the available values}

\item{top_n}{a numeric - an integer for selecting the top ranked samplesets}
}
\value{
a list with four character vectors (id,name,type,value)
}
\description{
get_biofeature_ranking()
}
\examples{
junk <- get_biofeature_ranking(samplesetIds = c("U2FtcGxlU2V0VHlwZTo0OTY2", "U2FtcGxlU2V0VHlwZToyNDgy", "U2FtcGxlU2V0VHlwZTo4NzQ="), top_n = 10)
}
