% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_biofeature_ranking}
\alias{get_biofeature_ranking}
\title{get_biofeature_ranking()}
\usage{
get_biofeature_ranking(
  compendium = "vespucci",
  samplesetNames = NULL,
  normalization = "legacy",
  rank = "uncentered_correlation",
  rankTarget = "biofeatures",
  top_n = 50,
  useIds = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{samplesetNames}{A vector of character strings (here sampleset names)}

\item{normalization}{A string ('legacy' as default)}

\item{rank}{A string ('magnitude' as default) - use \code{\link{get_ranking}}}

\item{rankTarget}{A string ('biofeature' as default)}

\item{top_n}{A numeric - an integer for selecting the top ranked samplesets}

\item{useIds}{A logical - TRUE as default}
}
\value{
A data.frame with three columns id, name, value
}
\description{
get_biofeature_ranking()
}
\examples{
\dontrun{
my_ss <- c("U2FtcGxlU2V0VHlwZToyMTk4Nw==", "U2FtcGxlU2V0VHlwZToyMTk4OA==",
"U2FtcGxlU2V0VHlwZToyMTk4OQ==", "U2FtcGxlU2V0VHlwZToyMTk5MA==")
get_biofeature_ranking(samplesetNames = my_ss, top_n = 10,
normalization = "tpm", useIds = TRUE)
}
}
