% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_compendium_data_source}
\alias{get_compendium_data_source}
\title{Get compendium data sources}
\usage{
get_compendium_data_source(compendium = "vitis_vinifera")
}
\arguments{
\item{compendium}{a character - the selected compendium}
}
\value{
a character vector containing the available data sources
}
\description{
Get compendium data sources
}
\examples{
get_compendium_data_source()
}
