% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_experiments}
\alias{get_experiments}
\title{Get available experiments information for the selected compendium,
use \code{\link{get_compendia}} to check all the available compendia}
\usage{
get_experiments(compendium = "vitis_vinifera")
}
\arguments{
\item{compendium}{a character - the selected compendium}
}
\value{
a data.frame
}
\description{
Get available experiments information for the selected compendium,
use \code{\link{get_compendia}} to check all the available compendia
}
\examples{
get_experiments()
}
