% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_experiments}
\alias{get_experiments}
\title{Get experiment Accessids and name
use \code{\link{get_available_compendia}} to check all the available compendia}
\usage{
get_experiments(compendium = "vespucci", sampleName = NULL)
}
\arguments{
\item{compendium}{string - the selected compendium}

\item{sampleName}{string - if NULL(default) returns all available experiments ids
for the selected compendium}
}
\value{
a data.frame with experimentAccessId and EsperimentName
}
\description{
Get experiment Accessids and name
use \code{\link{get_available_compendia}} to check all the available compendia
}
\examples{
get_experiments()
get_experiments(sampleName="GSM671721.ch1")
}
