% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_ontologies}
\alias{get_ontologies}
\title{Get all available ontologies for the selected compendium}
\usage{
get_ontologies(compendium = "vespucci")
}
\arguments{
\item{compendium}{A string - the selected compendium}
}
\value{
A data.frame
}
\description{
Get all available ontologies for the selected compendium
}
\examples{
get_ontologies()
}
