% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_ontology_structure}
\alias{get_ontology_structure}
\title{get_ontology_structure}
\usage{
get_ontology_structure(compendium = "vespucci", name = "Gene ontology")
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{name}{a character - the name of the ontolgy of interest}
}
\value{
a character
}
\description{
get_ontology_structure
}
\examples{
get_ontology_structure()
}
