% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_OntologyNode}
\alias{get_OntologyNode}
\title{get_OntologyNode}
\usage{
get_OntologyNode(
  compendium = "vespucci",
  normalization = "tpm",
  originalId = "PO_0009005",
  ontology_Name = "Plant ontology",
  descendantOf = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string ('legacy' as default)}

\item{originalId}{A string - the ontology term}

\item{ontology_Name}{A string - the name of the selected ontology}

\item{descendantOf}{A string - the id from the ontology term}
}
\value{
A data.frame
}
\description{
get_OntologyNode
}
\examples{
\dontrun{
root <- get_OntologyNode(normalization = "tpm", originalId = "PO_0009005",
ontology_Name = "Plant ontology")
root_terms <- get_OntologyNode(normalization = "tpm",
descendantOf = "T250b2xvZ3lOb2RlVHlwZToyMzg4Nzk0")
}
}
