% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_platform_information}
\alias{get_platform_information}
\title{Get information about all available platforms for the selected compendium,
use \code{\link{get_available_compendia}} to check all the available compendia}
\usage{
get_platform_information(compendium = "vespucci")
}
\arguments{
\item{compendium}{A string - the selected compendium}
}
\value{
A data.frame with five columns: accessId, name, description, source, type
}
\description{
Get information about all available platforms for the selected compendium,
use \code{\link{get_available_compendia}} to check all the available compendia
}
\examples{
info <- get_platform_information()
dplyr::count(info,type, source)
}
