% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_platform_types}
\alias{get_platform_types}
\title{get available platform types}
\usage{
get_platform_types(compendium = "vespucci")
}
\arguments{
\item{compendium}{A string - the selected compendium}
}
\value{
A vector of character strings
}
\description{
get available platform types
}
\examples{
get_platform_types()
}
