% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_ranking}
\alias{get_ranking}
\title{get_ranking}
\usage{
get_ranking(compendium = "vespucci")
}
\arguments{
\item{compendium}{a character - the selected compendium}
}
\value{
a list
}
\description{
get_ranking
}
\examples{
get_ranking()
}
