% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sample_by_gsm}
\alias{get_sample_by_gsm}
\title{get_sample_by_gsm}
\usage{
get_sample_by_gsm(compendium = "vespucci", sampleName_Icontains = "GSM1313535")
}
\arguments{
\item{compendium}{string - the selected compendium}

\item{sampleName_Icontains}{string - GSM (GEO Sample access id)}
}
\value{
a data.frame with three columns: id, sampleName, description
}
\description{
get_sample_by_gsm
}
\examples{
get_sample_by_gsm()
}
