% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sample_by_gsm}
\alias{get_sample_by_gsm}
\title{retrieve internal id, name (GSM), description for sample providing sampleName (GSM)}
\usage{
get_sample_by_gsm(compendium = "vespucci", sampleName_Icontains = "GSM1313535")
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{sampleName_Icontains}{A string - GSM (GEO Sample access id)}
}
\value{
A data.frame with three columns: sampleId, sampleName, sampleDescription
}
\description{
retrieve internal id, name (GSM), description for sample providing sampleName (GSM)
}
\examples{
\dontrun{
get_sample_by_gsm(sampleName_Icontains="GSM1313535")
}
}
