% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sample_info}
\alias{get_sample_info}
\title{get_sample_info}
\usage{
get_sample_info(compendium = "vitis_vinifera", sampleName = NULL)
}
\arguments{
\item{sampleName}{}
}
\value{

}
\description{
get_sample_info
}
