% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samples}
\alias{get_samples}
\title{Get all samples measured with a given platform}
\usage{
get_samples(compendium = "vespucci", platformAccessId = "GPL11004")
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{platformAccessId}{string}
}
\value{
A data.frame
}
\description{
Get all available samples for the selected compendium,
use \code{\link{get_available_compendia}} to check all the available compendia
}
\examples{
get_samples()
}
