% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samples_by_gse}
\alias{get_samples_by_gse}
\title{retrieve internal id, name (GSM), description from all samples with experiment id (GSE)}
\usage{
get_samples_by_gse(
  compendium = "vespucci",
  normalization = "legacy",
  experimentAccessId = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string - 'tpm' (as default), 'limma' or 'legacy'}

\item{experimentAccessId}{A string - GSE (GEO Series (experiment) access id)}
}
\value{
A data.frame with three columns: sampleId, sampleName, sampleDescription
}
\description{
retrieve internal id, name (GSM), description from all samples with experiment id (GSE)
}
\examples{
\dontrun{
get_samples_by_gse(experimentAccessId = "GSE98923")
}
}
