% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samples_by_gse}
\alias{get_samples_by_gse}
\title{get_samples_by_gse}
\usage{
get_samples_by_gse(
  compendium = "vespucci",
  experiment_ExperimentAccessId = "GSE54347"
)
}
\arguments{
\item{compendium}{A string- the selected compendium}

\item{experiment_ExperimentAccessId}{A string- GSE (GEO Series (experiment) access id)}
}
\value{
a data.frame with three columns id, name, description
}
\description{
get_samples_by_gse
}
\examples{
get_samples_by_gse()
}
