% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sampleset_id}
\alias{get_sampleset_id}
\title{get_sampleset_id}
\usage{
get_sampleset_id(compendium = "vespucci", name = NULL, samples = NULL)
}
\arguments{
\item{compendium}{a character - the selected compendium}

\item{name}{a character - the sampleset of interest}

\item{samples}{a character - unique id for the sampleset}
}
\value{
a data.frame eith two columns: id and name
}
\description{
get_sampleset_id
}
\examples{
get_sampleset_id(name = "GSE27180_48hours-1-vs-GSE27180_0h-2")
get_sampleset_id(sample ="U2FtcGxlVHlwZTox")
}
