% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sampleset_id}
\alias{get_sampleset_id}
\title{get_sampleset_id}
\usage{
get_sampleset_id(
  compendium = "vespucci",
  version = "latest",
  samples = NULL,
  name_In = NULL
)
}
\arguments{
\item{compendium}{A string- the selected compendium}

\item{version}{A string- either 'latest' or 'legacy'}

\item{name_In}{A vector of character strings - the sampleset name}

\item{name}{A string- the sampleset of interest}
}
\value{
a data.frame with two columns: id and name
}
\description{
get_sampleset_id
}
\examples{
get_sampleset_id(name_In = "GSE27180_48hours-1-vs-GSE27180_0h-2")
get_sampleset_id(samples =c("U2FtcGxlVHlwZTox"))
}
