% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sampleset_id}
\alias{get_sampleset_id}
\title{get_sampleset_id}
\usage{
get_sampleset_id(
  compendium = "vespucci",
  normalization = "legacy",
  name_In = NULL,
  useIds = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string ('legacy' as default)}

\item{name_In}{A vector of character strings - either samplesetNames or sampleSetIds}

\item{useIds}{A logical (FALSE as default) - It allows using sampleSetIds}
}
\value{
A data.frame with two columns: id and name
}
\description{
get_sampleset_id
}
\examples{
my_ss <- c("GSM671720.ch1-vs-GSM671719.ch1","GSM671721.ch1-vs-GSM671719.ch1"
,"GSM671722.ch1-vs-GSM671719.ch1","GSM147672.ch1-vs-GSM147690.ch1")
get_sampleset_id(name_In = my_ss, normalization = "legacy")
my_ids <- c("U2FtcGxlVHlwZToxMDI4","U2FtcGxlVHlwZTozMDQ5","U2FtcGxlVHlwZTo3MTY=")
get_sampleset_id(name_In = my_ids, normalization = "legacy", useIds = TRUE)
}
