% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samplesets_ranking}
\alias{get_samplesets_ranking}
\title{get_samplesets_ranking}
\usage{
get_samplesets_ranking(
  compendium = "vespucci",
  version = "legacy",
  rankTarget = "samplesets",
  rank = "magnitude",
  biofeaturesNames = NULL,
  biofeaturesIds = NULL,
  top_n = 50
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{version}{A string('legacy' as default)}

\item{rankTarget}{A string('sampleset' as default)}

\item{rank}{A string('magnitude' as default)}

\item{biofeaturesNames}{A vector of character strings (here gene_names)}

\item{biofeaturesIds}{A vector of character strings - the biofeature ids}

\item{top_n}{A numeric - an integer for selecting the top ranked samplesets}
}
\value{
A data.frame with four character vectors (id,name,type,value)
}
\description{
get_samplesets_ranking
}
\examples{
my_ids <- c("VIT_00s0332g00110","VIT_00s0332g00160","VIT_00s0396g00010","VIT_00s0505g00030")
get_samplesets_ranking(biofeaturesNames = my_ids, top_n = 10)
get_samplesets_ranking(biofeaturesIds=c("QmlvRmVhdHVyZVR5cGU6MQ==","QmlvRmVhdHVyZVR5cGU6Mg==",
 "QmlvRmVhdHVyZVR5cGU6Mw==","QmlvRmVhdHVyZVR5cGU6NA==","QmlvRmVhdHVyZVR5cGU6NQ=="), top_n = 10)
}
