% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_samplesets_ranking}
\alias{get_samplesets_ranking}
\title{get_samplesets_ranking}
\usage{
get_samplesets_ranking(
  compendium = "vespucci",
  normalization = "legacy",
  rankTarget = "samplesets",
  rank = "magnitude",
  biofeaturesNames = NULL,
  top_n = 50,
  useIds = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string ('legacy' as default)}

\item{rankTarget}{A string ('sampleset' as default)}

\item{rank}{A string ('magnitude' as default)}

\item{biofeaturesNames}{A vector of character strings (here gene_names)}

\item{top_n}{A numeric - an integer for selecting the top ranked samplesets}

\item{useIds}{A logical - TRUE as default}
}
\value{
A data.frame with three columns id, name, value
}
\description{
get_samplesets_ranking
}
\examples{
get_samplesets_ranking(biofeaturesNames=c("QmlvRmVhdHVyZVR5cGU6MQ==",
"QmlvRmVhdHVyZVR5cGU6Mg==","QmlvRmVhdHVyZVR5cGU6Mw==","QmlvRmVhdHVyZVR5cGU6NA==",
 "QmlvRmVhdHVyZVR5cGU6NQ=="), useIds = TRUE, top_n = 10)
}
