% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sparql_annotation_triples}
\alias{get_sparql_annotation_triples}
\title{get_sparql_annotation_triples}
\usage{
get_sparql_annotation_triples(
  compendium = "vespucci",
  target = "sample",
  query = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{target}{A string}

\item{query}{A string - sparql query}
}
\value{
A list
}
\description{
get_sparql_annotation_triples
}
\examples{
my_query <- "SELECT ?s ?p ?o WHERE { ?s <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://purl.obolibrary.org/obo/NCIT_C19157> . ?s <http://www.w3.org/1999/02/22-rdf-syntax-ns#type> <http://purl.obolibrary.org/obo/PO_0009010>}"
get_sparql_annotation_triples(query = my_query)
}
