% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_sparql_annotation_triples}
\alias{get_sparql_annotation_triples}
\title{get_sparql_annotation_triples}
\usage{
get_sparql_annotation_triples(
  compendium = "vespucci",
  normalization = "tpm",
  target = NULL,
  query = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string - 'tpm' (as default), 'limma' or 'legacy'}

\item{target}{A string - either 'sample' or 'biofeature'}

\item{query}{A string - sparql query}
}
\value{
A data.frame with 3 columns
}
\description{
get_sparql_annotation_triples
}
\examples{
\dontrun{
my_query=paste0('SELECT ?s ?p ?o WHERE { ?s <http://www.w3.org/1999/02/22-rdf-syntax-ns#type>',
'<http://purl.obolibrary.org/obo/NCIT_C19157>',
'. ?s <http://www.w3.org/1999/02/22-rdf-syntax-ns#type>',
'<http://purl.obolibrary.org/obo/PO_0009010>}')
get_sparql_annotation_triples(normalization = "legacy", target = "sample", query = my_query)
}
}
