% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{merge_modules}
\alias{merge_modules}
\title{Union of two modules}
\usage{
merge_modules(mod1, mod2)
}
\arguments{
\item{mod1}{first module to merge}

\item{mod2}{second module to merge}
}
\value{
A matrix - the module
}
\description{
Union of two modules
}
\examples{
\dontrun{
my_ss <- c("GSM671720.ch1-vs-GSM671719.ch1","GSM671721.ch1-vs-GSM671719.ch1"
,"GSM671722.ch1-vs-GSM671719.ch1","GSM147672.ch1-vs-GSM147690.ch1")
mod_ss <- create_module_ss(samplesetNames = my_ss, normalization = "legacy")
my_bf <- c("VIT_00s0246g00220","VIT_00s0332g00060","VIT_00s0332g00110"
,"VIT_00s0332g00160","VIT_00s0396g00010","VIT_00s0505g00030","VIT_00s0505g00060"
,"VIT_00s0873g00020","VIT_00s0904g00010")
mod_bf <- create_module_bf(biofeaturesNames= my_bf, normalization = "legacy")
mod_union <- merge_modules(mod1 = mod_ss, mod2 = mod_bf)
}
}
