% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  compendium = "vespucci",
  normalization = NULL,
  biofeaturesNames = NULL,
  samplesetNames = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{normalization}{A string - either 'limma','tpm_sample' or legacy as normalization}

\item{biofeaturesNames}{A character vector (here gene_names)}

\item{samplesetNames}{A character vector - the sampleSets names}
}
\value{
An html page with the heatmap
}
\description{
plot_heatmap
}
\examples{
\dontrun{
my_plot_html <- plot_heatmap(QuickSearch_ABAcoreset)
tempDir <- tempfile()
dir.create(tempDir)
htmlFile <- file.path(tempDir, "plot_heatmap.html")
xml2::write_html(my_plot_html,file=htmlFile)
rstudioapi::viewer(htmlFile)
}
}
