% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  compendium = "vespucci",
  version = "legacy",
  biofeaturesNames = NULL,
  samplesetNames = NULL
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{version}{A string ('legacy' as default)}

\item{biofeaturesNames}{A character vector (here gene_names)}

\item{samplesetNames}{A character vector - the sampleSets names}
}
\value{
An html page with the heatmap
}
\description{
plot_heatmap
}
