% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_module_heatmap}
\alias{plot_module_heatmap}
\title{plot heatmap from a module}
\usage{
plot_module_heatmap(
  compendium = "vespucci",
  module = NULL,
  normalization = "tpm",
  type = "json",
  plot = TRUE,
  alternativeColoring = TRUE,
  min = -6,
  max = 6,
  sorted = FALSE
)
}
\arguments{
\item{compendium}{A string - the selected compendium}

\item{module}{A matrix with valid rownames (biofeatureNames) and colnames (samplesetsNames)}

\item{normalization}{A string - either 'limma','tpm' or legacy as normalization}

\item{type}{A string -  either 'html'  or 'json}

\item{plot}{A logical - it returns the graphics object}

\item{alternativeColoring}{A logical - if TRUE (defautl) a color blind friendly palette is used}

\item{min}{A numeric (-6 default)}

\item{max}{A numeric (6 default)}

\item{sorted}{A logical - it returns sorted index for both bf and ss}
}
\value{
Either a json, an html, a plotly htmlwidget, or a list of sorted features
}
\description{
plot heatmap from a module
}
\examples{
\dontrun{
gene_names <- get_biofeature_id(id_In = c('VIT_00s0246g00220',
'VIT_00s0332g00060','VIT_00s0332g00110'), useIds = F)
mod <- create_module(biofeaturesNames=gene_names$id, normalization = "tpm", useIds = T)
plot_module_heatmap(module = mod, normalization = "tpm", plot = TRUE)
sorted_idx <- plot_module_heatmap(mod = mod, sorted = TRUE, plot = FALSE)
}
}
