% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_network_coexpression}
\alias{plot_network_coexpression}
\title{plot_network_coexpression}
\usage{
plot_network_coexpression(
  compendium = "vespucci",
  version = "legacy",
  biofeaturesNames = NULL,
  samplesetNames = NULL
)
}
\arguments{
\item{compendium}{string - the selected compendium}

\item{version}{string ('legacy' as default)}

\item{samplesetNames}{a character vector - the sampleSets names}

\item{biofeatureNames}{a character vector (here gene_names)}
}
\value{
an html page with the coexpression network
}
\description{
plot_network_coexpression
}
