% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{view_plot}
\alias{view_plot}
\title{Show a plot using the rstudio viewer}
\usage{
view_plot(module = NULL, type = "network", threshold = 0.7)
}
\arguments{
\item{module}{A matrix with valid rownames (biofeatureNames) and colnames (samplesetsNames)}

\item{type}{A string -  either 'network', 'heatmap' or 'distribution'}

\item{threshold}{A numeric - A Pearson correalation value}
}
\value{
A plotly htmlwidget
}
\description{
Show a plot using the rstudio viewer
}
